PR #352 review: required isViable param + extract buildNodeData
- `getNextOrPrevious`: drop the optional `?` on `isViable`. The only no-
  predicate callers were tests; production (`useCommon`) always supplies
  one. Tests now pass `acceptAll = () => true` for the pure structural
  walk. The "absence of the predicate is byte-equivalent" case is gone —
  trivially true under a required predicate.
- `buildLeafNodeData` (inline in `keyboard.ts`) was a near-duplicate of
  `buildNodeData` in `JsonEditor.tsx`. Extracted the canonical builder
  to `src/utils/buildNodeData.ts` and reuse from both call sites.
  `rootName` defaults to `''` for callers that never target the root.
