+++ src/utils/keyboard.ts
@@ -9,6 +9,7 @@ import {
   type SortFunction,
   type TabDirection,
 } from '../types'
+import { buildNodeData } from './buildNodeData'
 import { extract } from './extract'
 import { isCollection } from './misc'
 
@@ -155,13 +156,15 @@ export const getNextOrPrevious = (
   path: CollectionKey[],
   nextOrPrev: TabDirection = 'next',
   sort: SortFunction,
-  // Optional viability predicate. When supplied, candidate leaves whose
-  // synthesized `NodeData` fails the predicate are skipped: the function
-  // recurses with the failed candidate as the new starting point,
-  // continuing in the same direction until a viable leaf or `null` is
-  // reached. Lets Tab navigation skip filtered-out or non-editable nodes
-  // up front instead of relying on a downstream redirect to bounce them.
-  isViable?: (nodeData: NodeData) => boolean
+  // Viability predicate. Candidate leaves whose synthesized `NodeData`
+  // fails the predicate are skipped: the function recurses with the
+  // failed candidate as the new starting point, continuing in the same
+  // direction until a viable leaf or `null` is reached. Lets Tab
+  // navigation skip filtered-out or non-editable nodes up front instead
+  // of relying on a downstream redirect to bounce them. Pass `() => true`
+  // for a pure structural walk (tests; never in production — the editor
+  // always supplies a real predicate via `useCommon`).
+  isViable: (nodeData: NodeData) => boolean
 ): CollectionKey[] | null => {
   const parentPath = path.slice(0, path.length - 1)
   const thisKey = path.slice(-1)[0]
@@ -195,45 +198,10 @@ export const getNextOrPrevious = (
   }
   if (!candidate) return null
 
-  if (isViable && !isViable(buildLeafNodeData(fullData, candidate, sort))) {
+  if (!isViable(buildNodeData(fullData, candidate, '', sort))) {
     return getNextOrPrevious(fullData, candidate, nextOrPrev, sort, isViable)
   }
   return candidate
-}
-
-// Synthesizes the `NodeData` shape for an arbitrary leaf path — used to
-// feed the `isViable` predicate above. Mirrors `buildNodeData` in
-// `JsonEditor.tsx` (kept inline to keep this util free of cross-file
-// dependencies on the editor). `index` derivation matches the renderer's
-// sort-aware semantics so a custom `searchFilter` / `allowEdit` sees the
-// same `index` it would during render.
-const buildLeafNodeData = (
-  fullData: JsonData,
-  path: CollectionKey[],
-  sort: SortFunction
-): NodeData => {
-  const parentPath = path.slice(0, -1)
-  const parentData = (extract(fullData, parentPath) ?? null) as object | null
-  const key = path[path.length - 1]
-  const value = extract(fullData, path) as JsonData
-  let index = 0
-  if (Array.isArray(parentData)) {
-    index = typeof key === 'number' ? key : Number(key)
-  } else if (parentData) {
-    const entries = Object.entries(parentData) as Array<[string | number, unknown]>
-    sort(entries, (entry) => entry)
-    index = entries.findIndex(([k]) => k === key)
-  }
-  return {
-    key,
-    path,
-    level: path.length,
-    index,
-    value,
-    size: isCollection(value) ? Object.keys(value as object).length : null,
-    parentData,
-    fullData,
-  }
 }
 
 // If the node at "path" is a collection, tries the first/last child of that
